finished 2nd code review

+++ app.py
@@ -1,8 +1,8 @@
 """Notes application."""
 
-from flask import Flask, request, redirect, render_template, session, flash
+from flask import Flask, redirect, render_template, session, flash
 from models import db, connect_db, User, Note
-from forms import RegisterForm, LoginForm, CSRFProtectForm, AddNoteForm
+from forms import RegisterForm, LoginForm, CSRFProtectForm, AddNoteForm, EditNoteForm
 
 app = Flask(__name__)
 app.config['SQLALCHEMY_DATABASE_URI'] = 'postgresql:///notes'
@@ -19,7 +19,10 @@ debug = DebugToolbarExtension(app)
 @app.get("/")
 def redirect_register():
     """Redirect to /register"""
-
+    # reroute depending on whether you're logged in or not
+    if session.get("user_id"):
+        username = session.get("user_id")
+        return redirect(f"/users/{username}")
     return redirect("/register")
 
 @app.route("/register", methods=["GET", "POST"])
@@ -98,42 +101,91 @@ def logout_user():
 
 @app.post("/users/<username>/delete")
 def delete_user(username):
-    """Delete the user"""
+    """Delete the current user."""
 
     form = CSRFProtectForm()
 
-    if form.validate_on_submit():
-        user = User.query.get_or_404(username)
-        notes = user.notes
+    if session.get("user_id") and session.get("user_id") == username:
+        if form.validate_on_submit():
+            user = User.query.get_or_404(username)
 
-        for note in notes:
-            db.session.delete(note)
+            Note.query.filter(Note.owner == username).delete()
 
-        db.session.commit()
+            db.session.commit()
 
-        db.session.delete(user)
-        db.session.commit()
+            db.session.delete(user)
+            db.session.commit()
 
-        session.pop("user_id", None)
+            session.pop("user_id", None)
 
-        return redirect('/')
+            return redirect('/')
+
+    return redirect('/')
 
 @app.route("/users/<username>/notes/add", methods=["GET", "POST"])
 def add_note(username):
-    """Add note"""
+    """Add a note for the current user."""
 
     form = AddNoteForm()
 
-    if form.validate_on_submit():
-        title = form.title.data
-        content = form.content.data
+    if session.get("user_id") and session.get("user_id") == username:
+        if form.validate_on_submit():
+            title = form.title.data
+            content = form.content.data
 
-        note = Note(title=title, content=content, owner=username)
+            note = Note(title=title, content=content, owner=username)
 
-        db.session.add(note)
-        db.session.commit()
+            db.session.add(note)
+            db.session.commit()
 
-        return redirect(f'/users/{username}')
+            return redirect(f'/users/{username}')
 
-    else:
-        return render_template("note.html", form=form)
+        else:
+            return render_template("add-note.html", form=form)
+
+    return redirect('/')
+
+@app.route("/notes/<int:note_id>/update", methods=["GET", "POST"])
+def update_note(note_id):
+    """Update a specified note."""
+    note = Note.query.get_or_404(note_id)
+    username = note.owner
+
+    form = EditNoteForm(obj=note)
+
+    if session.get("user_id") and session.get("user_id") == username:
+        if form.validate_on_submit():
+            title = form.title.data
+            content = form.content.data
+
+            note.title = title
+            note.content = content
+
+            db.session.add(note)
+            db.session.commit()
+
+            username = note.user.username
+
+            return redirect(f'/users/{username}')
+
+        else:
+            return render_template("edit-note.html", form=form)
+
+    return redirect('/')
+
+@app.post("/notes/<int:note_id>/delete")
+def delete_note(note_id):
+    """Delete a specified note."""
+
+    form = CSRFProtectForm()
+    note = Note.query.get_or_404(note_id)
+    username = note.owner
+
+    if session.get("user_id") and session.get("user_id") == username:
+        if form.validate_on_submit():
+            db.session.delete(note)
+            db.session.commit()
+
+            return redirect(f'/users/{username}')
+
+        return redirect("/")
